+++ src/features/registration/profile-setup-page.tsx
